# api/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for managing climbing board routes and data. Features cursor-based pagination, grade information for multiple board angles, and comprehensive filtering capabilities.",
        "title": "Kilter Together API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "",
    "basePath": "/api",
    "paths": {
        "/boards": {
            "get": {
                "description": "Retrieve all available board/product size options for filtering climbs. Returns a list of board configurations with their IDs and human-readable names.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "boards"
                ],
                "summary": "Get available board options",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved board options",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.BoardOption"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/climbs": {
            "get": {
                "description": "Retrieve a paginated list of climbing routes with optional filtering. Uses cursor-based pagination for efficient navigation through large datasets.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "climbs"
                ],
                "summary": "Get paginated climbs",
                "parameters": [
                    {
                        "type": "string",
                        "example": "2025-05-24 04:07:17.406545",
                        "description": "Pagination cursor (timestamp for next page)",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "example": 10,
                        "description": "Number of items per page (1-100)",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "swooped",
                        "description": "Filter climbs by name (partial match)",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "Filter climbs by board/product size ID",
                        "name": "board_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved climbs",
                        "schema": {
                            "$ref": "#/definitions/models.CursorPaginatedClimbsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/images/{filename}": {
            "get": {
                "description": "Serve static image files for board layouts and holds. Images are used to display the physical board layout with hold positions for each climb.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "image/png",
                    "image/jpeg",
                    "image/gif"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Serve board layout images",
                "parameters": [
                    {
                        "type": "string",
                        "example": "original-16x12-bolt-ons-v2.png",
                        "description": "Image filename",
                        "name": "filename",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully served image file",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad request - invalid filename",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not found - image file does not exist",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.BoardOption": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Unique board identifier",
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "description": "Human-readable board name",
                    "type": "string",
                    "example": "Original 12x12"
                }
            }
        },
        "models.Climb": {
            "type": "object",
            "properties": {
                "climb_name": {
                    "description": "Name/title of the climb",
                    "type": "string",
                    "example": "swooped"
                },
                "created_at": {
                    "description": "Creation timestamp",
                    "type": "string",
                    "example": "2018-12-06 21:15:01.127371"
                },
                "description": {
                    "description": "Optional description",
                    "type": "string",
                    "example": "A challenging overhang problem"
                },
                "frames": {
                    "description": "Hold positions and rotations",
                    "type": "string",
                    "example": "p1080r15p1110r15p1131r12"
                },
                "grades": {
                    "description": "Map of angle to grade info",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/models.GradeInfo"
                    }
                },
                "image_filenames": {
                    "description": "JSON array string of image filenames",
                    "type": "string",
                    "example": "[\"layout1.png\",\"layout2.png\"]"
                },
                "product_size_id": {
                    "description": "Board/product size identifier",
                    "type": "integer",
                    "example": 1
                },
                "setter_name": {
                    "description": "Username of the climb setter",
                    "type": "string",
                    "example": "jwebxl"
                },
                "uuid": {
                    "description": "Unique identifier for the climb",
                    "type": "string",
                    "example": "F01419E12672459396CA62E3655ABC46"
                }
            }
        },
        "models.CursorPaginatedClimbsResponse": {
            "type": "object",
            "properties": {
                "climbs": {
                    "description": "Array of climb objects",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Climb"
                    }
                },
                "has_more": {
                    "description": "Whether more pages are available",
                    "type": "boolean",
                    "example": true
                },
                "next_cursor": {
                    "description": "Cursor for next page (timestamp)",
                    "type": "string",
                    "example": "2025-05-24 04:07:17.406545"
                },
                "page_size": {
                    "description": "Number of items per page",
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "models.GradeInfo": {
            "type": "object",
            "properties": {
                "boulder": {
                    "description": "Boulder grade in format \"grade/V-scale\"",
                    "type": "string",
                    "example": "7a/V6"
                },
                "route": {
                    "description": "Route grade in format \"grade/YDS\"",
                    "type": "string",
                    "example": "7c/5.12d"
                }
            }
        }
    },
    "externalDocs": {
        "description": "Kilter Together API Documentation",
        "url": "https://github.com/lczm/boardbuddy/blob/main/api/API.md"
    }
}